Vanilla Bugfix: The Ruins of Alph research center's roof color at night looks wrong
The dungeons' map group mostly has indoor maps that don't need roof colors, but maps/RuinsOfAlphOutside.blk is an exception. It appears to have poorly-chosen roof colors: the morning/day colors are the same default gray as the unused group 0, and the night colors combine the light default gray and the dark red of Cinnabar's night roofs.