Show the tops of leaders heads on the trainer card
All the Gym Leaders' names in Japanese were short enough to appear above their heads on the trainer card.
That wouldn't work in English, so the names were erased from the trainer card graphics, but the Gym Leaders' heads were left with their tops missing. Fortunately, this is easy to fix.