Expand tilesets from 192 to 255 tiles
By default, tilesets in Gen 2 are limited to 192 tiles. But we can see with BGB's VRAM viewer that there's room for more.
VRAM is divided into six areas, each 128 tiles large. The bottom-left and bottom-right areas are mostly for map tiles: $00–$5F on the left, $80–$DF on the right. The middle-left area ($80–$FF), and the bottom two rows of the bottom-left area ($60–$7F), are for text character tiles. The bottom two rows of the bottom-right area ($E0–$FF) are unused. Furthermore, many of the text tiles are unused as well, including the ümläuted letters, most of the bold letters, etc.

It's fairly simple to use $E0–$FF for 32 more map tiles. We can also use $60–$7E by moving its necessary text tiles into the main font area, leaving only $7F for the space character and allowing 255 map tiles.

(The code for this feature was adapted from Pokémon Polished Crystal.)

Co-Authored-By: Rangi <35663410+Rangi42@users.noreply.github.com>