Allow tiles to have different attributes in different blocks (including X and Y flip)
Maps in pokecrystal are designed with blocks, not tiles, where each block (aka "metatile") is a 4x4 square of tiles. Each 8x8-pixel tile always has the same appearance in every block. But the GameBoy hardware is capable of more. The same tile graphic can be reused with different attributes—not just color, but X and Y flip, as well as "priority" to appear above sprites.l.)

(The code for this feature was adapted from Pokémon Polished Crystal.)

Co-Authored-By: Rangi <35663410+Rangi42@users.noreply.github.com>